Split off functions into module
