+++ src/player_control/camera/rig.rs
@@ -1,13 +1,15 @@
 use crate::file_system_interaction::config::GameConfig;
 use crate::player_control::actions::CameraAction;
+use crate::player_control::camera::rig::arm::{get_arm_distance, get_zoom_smoothness, set_arm};
 use crate::player_control::camera::{IngameCamera, IngameCameraKind};
-use crate::util::smoothness_to_lerp_factor;
-use crate::util::trait_extension::{F32Ext, Vec2Ext};
+use crate::util::trait_extension::Vec2Ext;
 use anyhow::{Context, Result};
 use bevy::prelude::*;
 use bevy_dolly::prelude::*;
 use bevy_rapier3d::prelude::*;
 use leafwing_input_manager::prelude::ActionState;
+
+mod arm;
 
 pub fn update_rig(
     time: Res<Time>,
@@ -27,7 +29,7 @@ pub fn update_rig(
         if camera.kind == IngameCameraKind::FixedAngle {
             let yaw_pitch = rig.driver_mut::<YawPitch>();
             yaw_pitch.yaw_degrees = 0.;
-            yaw_pitch.pitch_degrees = -80.0;
+            yaw_pitch.pitch_degrees = config.camera.fixed_angle.pitch;
         } else {
             let camera_movement = get_camera_movement(actions)?;
             if !camera_movement.is_approx_zero() {
@@ -47,24 +49,6 @@ pub fn update_rig(
     Ok(())
 }
 
-fn get_arm_distance(
-    camera: &IngameCamera,
-    transform: &Transform,
-    rapier_context: &RapierContext,
-    config: &GameConfig,
-) -> Option<f32> {
-    match camera.kind {
-        IngameCameraKind::ThirdPerson => Some(get_distance_to_collision(
-            rapier_context,
-            config,
-            camera,
-            transform,
-        )),
-        IngameCameraKind::FixedAngle => Some(camera.desired_distance),
-        _ => None,
-    }
-}
-
 fn get_camera_movement(actions: &ActionState<CameraAction>) -> Result<Vec2> {
     actions
         .axis_pair(CameraAction::Orbit)
@@ -77,10 +61,8 @@ fn set_yaw_pitch(rig: &mut Rig, camera: &IngameCamera, camera_movement: Vec2, co
     let yaw = -camera_movement.x * config.camera.mouse_sensitivity_x;
     let pitch = -camera_movement.y * config.camera.mouse_sensitivity_y;
     yaw_pitch.rotate_yaw_pitch(yaw.to_degrees(), pitch.to_degrees());
-    let (most_acute_looking_down, most_acute_looking_up) = get_most_acute_degrees(config, camera);
-    yaw_pitch.pitch_degrees = yaw_pitch
-        .pitch_degrees
-        .clamp(most_acute_looking_up, most_acute_looking_down);
+    let (min_pitch, max_pitch) = get_pitch_extrema(config, camera);
+    yaw_pitch.pitch_degrees = yaw_pitch.pitch_degrees.clamp(min_pitch, max_pitch);
 }
 
 fn set_look_at(rig: &mut Rig, camera: &IngameCamera) {
@@ -102,18 +84,38 @@ fn set_position(rig: &mut Rig, camera: &IngameCamera) {
     rig.driver_mut::<Position>().position = target;
 }
 
-fn get_most_acute_degrees(config: &GameConfig, camera: &IngameCamera) -> (f32, f32) {
+fn get_pitch_extrema(config: &GameConfig, camera: &IngameCamera) -> (f32, f32) {
     match camera.kind {
         IngameCameraKind::ThirdPerson => (
-            config.camera.third_person.max_pitch,
             config.camera.third_person.min_pitch,
+            config.camera.third_person.max_pitch,
         ),
         IngameCameraKind::FirstPerson => (
-            config.camera.first_person.max_pitch,
             config.camera.first_person.min_pitch,
+            config.camera.first_person.max_pitch,
         ),
         _ => unreachable!(),
     }
+}
+
+fn set_desired_distance(
+    camera: &mut IngameCamera,
+    actions: &ActionState<CameraAction>,
+    config: &GameConfig,
+) {
+    let zoom = actions.clamped_value(CameraAction::Zoom) * config.camera.third_person.zoom_speed;
+    let (min_distance, max_distance) = match camera.kind {
+        IngameCameraKind::ThirdPerson => (
+            config.camera.third_person.min_distance,
+            config.camera.third_person.max_distance,
+        ),
+        IngameCameraKind::FixedAngle => (
+            config.camera.fixed_angle.min_distance,
+            config.camera.fixed_angle.max_distance,
+        ),
+        IngameCameraKind::FirstPerson => (0.0, 0.0),
+    };
+    camera.desired_distance = (camera.desired_distance - zoom).clamp(min_distance, max_distance);
 }
 
 fn set_smoothness(rig: &mut Rig, config: &GameConfig, camera: &IngameCamera) {
@@ -141,106 +143,4 @@ fn set_smoothness(rig: &mut Rig, config: &GameConfig, camera: &IngameCamera) {
                 config.camera.fixed_angle.rotation_smoothing;
         }
     }
-}
-
-fn get_zoom_smoothness(
-    config: &GameConfig,
-    camera: &IngameCamera,
-    rig: &Rig,
-    new_distance: f32,
-) -> f32 {
-    let current_distance = rig.driver::<Arm>().offset.z;
-    if new_distance < current_distance - 1e-4 {
-        match camera.kind {
-            IngameCameraKind::ThirdPerson => config.camera.third_person.zoom_in_smoothing,
-            IngameCameraKind::FixedAngle => config.camera.fixed_angle.zoom_in_smoothing,
-            _ => unreachable!(),
-        }
-    } else {
-        match camera.kind {
-            IngameCameraKind::ThirdPerson => config.camera.third_person.zoom_out_smoothing,
-            IngameCameraKind::FixedAngle => config.camera.fixed_angle.zoom_out_smoothing,
-            _ => unreachable!(),
-        }
-    }
-}
-
-fn set_arm(rig: &mut Rig, distance: f32, zoom_smoothness: f32, dt: f32) {
-    let factor = smoothness_to_lerp_factor(zoom_smoothness, dt);
-    let arm_length = &mut rig.driver_mut::<Arm>().offset.z;
-    *arm_length = arm_length.lerp(distance, factor);
-}
-
-fn set_desired_distance(
-    camera: &mut IngameCamera,
-    actions: &ActionState<CameraAction>,
-    config: &GameConfig,
-) {
-    let zoom = actions.clamped_value(CameraAction::Zoom) * config.camera.third_person.zoom_speed;
-    let (min_distance, max_distance) = match camera.kind {
-        IngameCameraKind::ThirdPerson => (
-            config.camera.third_person.min_distance,
-            config.camera.third_person.max_distance,
-        ),
-        IngameCameraKind::FixedAngle => (
-            config.camera.fixed_angle.min_distance,
-            config.camera.fixed_angle.max_distance,
-        ),
-        IngameCameraKind::FirstPerson => (0.0, 0.0),
-    };
-    camera.desired_distance = (camera.desired_distance - zoom).clamp(min_distance, max_distance);
-}
-
-fn get_distance_to_collision(
-    rapier_context: &RapierContext,
-    config: &GameConfig,
-    camera: &IngameCamera,
-    camera_transform: &Transform,
-) -> f32 {
-    let origin = camera.target.translation;
-    let direction = camera_transform.back();
-
-    let max_toi = camera.desired_distance;
-    let solid = true;
-    let mut filter = QueryFilter::only_fixed();
-    filter.flags |= QueryFilterFlags::EXCLUDE_SENSORS;
-
-    let min_distance = match camera.kind {
-        IngameCameraKind::ThirdPerson => config.camera.third_person.min_distance_to_objects,
-        _ => unreachable!(),
-    };
-
-    rapier_context
-        .cast_ray_and_get_normal(origin, direction, max_toi, solid, filter)
-        .map(|(_entity, ray_intersection)| {
-            get_distance_such_that_min_distance_from_collision_is_ensured(
-                ray_intersection,
-                direction,
-                min_distance,
-            )
-        })
-        .unwrap_or(max_toi)
-}
-
-fn get_distance_such_that_min_distance_from_collision_is_ensured(
-    ray_intersection: RayIntersection,
-    direction: Vec3,
-    min_distance: f32,
-) -> f32 {
-    //  Wall
-    //  ↑
-    //  │╲
-    //  │  ╲
-    //  │    ╲ Hypotenuse, direction = direction
-    //  │      ╲
-    //  │        ╲
-    //  │      ( α ╲
-    //  └─────────────→ Normal, magnitude = min_distance
-    //  │              ╲
-    //  │                ╲   Target vector. Magnitude = total length (toi) - hypotenuse
-    //  │                  ╲
-    let adjacent_side = min_distance;
-    let angle = direction.angle_between(-ray_intersection.normal);
-    let hypotenuse = adjacent_side / angle.cos();
-    ray_intersection.toi - hypotenuse
 }
